data-service/requirements.txt
Flask==2.0.1
kafka-python==2.0.2
SQLAlchemy==1.4.23